(api-cat) Move QueryResponse class to own package
