+++ applications/api-cat/src/main/java/no/acat/query/AcatQueryController.java
@@ -3,13 +3,12 @@ package no.acat.query;
 import io.swagger.annotations.ApiOperation;
 import io.swagger.annotations.ApiParam;
 import no.acat.model.ApiDocument;
-import no.acat.model.QueryResponse;
+import no.acat.model.queryresponse.QueryResponse;
 import no.acat.service.ElasticsearchService;
 import org.elasticsearch.action.search.SearchRequestBuilder;
 import org.elasticsearch.action.search.SearchResponse;
 import org.elasticsearch.index.query.QueryBuilder;
 import org.elasticsearch.index.query.QueryBuilders;
-import org.elasticsearch.search.SearchHit;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
